Project working. Need to redo the CRC mechanism. Need to add comments for doxygen.

+++ includes/JSON_manager.h
@@ -27,9 +27,9 @@ private:
     serializer ser;
     checksum_manager CRC;
 public:
-    uint8_t read_file(string filename);
+    unsigned int read_file(string filename);
     string get_json_config_string();
-    uint8_t parseToStructandSerialize(string filename);
+    unsigned int parseToStructandSerialize(string filename);
 };
